refactor(graph): extract store sessions
